Implement that V99x8 does not support mixed modes
At the same time, at least for some bogus combinations, the test of what
happens will apply too for the bitmap converter, so I implemented it.
But not all combinations were tested, so this is probably not accurate
yet (but already better than before).

For the M1-M3 bits it is at least clear that the behaviour for all these
modes is identical (including stopped refreshing of the VRAM, which is
not emulated). Also in combination with M4 and M5 (so only tested a few
of them) I saw the same behaviour.