[1/3] Scheduler optimizations, pendingSyncPoint()
In performance profile reports the Scheduler code is a bit sneaky: the cost of
the Scheduler is spread over several functions, each function only taking
around 1%, though if you add them all together you get roughly 5% of the total
time. In the past this was likely much less, but since the more accurate CPU
VRAM-access emulation we schedule each access, and this puts a lot more stress
on the Scheduler. This patch series is an attempt to speed it up.

As mentioned in the previous paragraph, each CPU-VRAM-access is now scheduled.
Before each such schedule operation we check whether there already is such an
access scheduled. This is done via the Scheduler::pendingSyncPoint() method.
Internally that method walks over all scheduled synchronization points (I
measured that typically there are between 6 and 8 pending sync points). This
patch replaces that call with a boolean flag (that is kept up to date). This
is redundant information, but it's a lot faster.

This patch is responsible for roughly half of the performance gain of this
series.