[3/3] Scheduler optimizations, SchedulerQueue
In openMSX the Scheduler works like this:
- MSX devices can register to execute some code when emulation reaches a
  certain point in time. These are called synchronization points.
- During CPU emulation the CPU and the Scheduler code cooperate to execute
  the registered code as soon as that time is reached. So synchronization
  points execute in order, from smaller to bigger timestamps.
- Synchronization points can be canceled.

So for an efficient Scheduler we need to be able to quickly:
- Get the synchronization point with the smallest time.
- Remove that smallest synchronization point (and get the 2nd smallest, etc).
- Insert new synchronization points at an arbitrary time (typically these won't
  be too far in the future).
- (Less important) cancel existing synchronization points.

A theoretical nice data structure for this (with good complexity guarantees) is
a priority queue (or sometimes called 'heap'). Though because our queue always
contains very few elements (typically 7, rarely more than 10), big-O complexity
guarantees are not the most important. Originally we did use a priority queue,
but a long time ago we already switched to a simple sorted vector (revision
908be42a).

Something I only recently observed (but obvious once you know it) is the
following: Synchronization points are often (more or less) periodic. I mean
sync points that are scheduled in the near future will, when they trigger,
often get rescheduled in the near future. And sync points for the distant
future are typically rescheduled for the distant future. So this means that
sync-point-inserts near the front of the queue are more frequent than inserts
near the end. In openMSX this is especially true for the CPU-VRAM access sync
points.

Before this patch the queue was implemented as a sorted vector with the
smallest sync point in front. This means that we frequently have inserts near
the front of the vector, and this requires copying most of the (bigger)
elements. An even more frequent operation is removing the smallest element, and
this also requires copying all the remaining elements.

One solution would be to reverse the sorting order. Though an even better
solution is to use a custom container that acts like a sorted vector but
with spare capacity both at the front and at the back (vector only has spare
capacity at the back). So that's what this patch does.

When inserting a new sync point we have to search for the right insertion point
(so that the elements remain sorted). Before we used binary search. But because
there are so few elements we now use linear search (with a sentinel).

--

Benchmarks for this full patch series show about 3x-5x less time is spend in
Scheduler related code! Resulting in between 2%-4% faster total emulation!