Don't zero-fill buffer for single channel sound devices
All SoundDevices must implement the following method:
   void generateChannels(int** buffers, unsigned num);
For each channel you get a buffer and you must fill those buffers with 'num'
number of samples. (There is an optimization for silent channels, but I'll
ignore that here). Separate buffers per channel allow features like per-channel
recording/muting/balance.

Though most of the time we don't actually need this per-channel stuff. And
those separate buffers would greatly increase the memory footprint (and
decrease CPU memory caching efficiency). So as a solution the caller can pass
the same buffer pointer for all the channels (or for all the channels that
don't need per-channel stuff). But for this to work the generateChannels()
method cannot simply overwrite the data in the given buffers, but it must add
the generated channel data to the existing data in the buffer. And this
requires that the (shared) buffer must initially be zero-filled.

I noticed that in some of my benchmarks slightly over 1% of total emulation
time was spend in this zero-filling of the (shared) buffer, this patch is an
attempt to reduce that cost.

--

A non-expanded MSX1/2 machine has the following sound devices:
 - PSG (3 channels)
 - key-click (1 channel)
 - cassette-player (1 channel)
2 of the 3 sound devices are single channel. For single channel devices we
don't actually need zero-fill/add (there is no opportunity to share buffers).
We can simply replace the content of the buffer with the generated data. So
that's what this patch does.

Next to key-click and cassette-player we also have the following single channel
sound devices:
 - SamplePlayer
 - VLM5030
 - LaserdiscPlayer
 - and some more devices based on DACSound16 (like key-click)

After this change I indeed see 3x less zero-fill calls, but surprisingly the
time spend in this routine only dropped by a factor 1.2x (not 3x). Part of this
can be explained by the higher native frequency of the PSG (so a bigger buffer
to fill). For a more complete explanation we also need to take cache effects
into account: the buffer is reused among all sound devices (actually it's
located on the stack and roughly the same stack memory location is used for all
devices). The first zero-fill call brings those memory locations from main RAM
(or L2?) into the CPU L1 cache (relatively expensive), but for the later calls
the buffer is already/still in L1 cache, so those calls are a lot cheaper (and
removing them doesn't gain as much). Anyway, the speedup is still measurable,
so I'm keeping it.

--

I also did a failed experiment. I'll only describe it briefly:

Very often one or more of the sound devices are silent (it's indeed very rare
to have all of key-click, cassetteplayer and PSG produce sound at the same
time). In that case the zero-filled buffer (the experiment was done before this
patch) remains zero-filled after the call to generateChannels(). I tried to
reuse those zeros for the next sound device (so skip the next zero-fill).

What I saw was, as expected, less time spend zero-filling, but the method
BlipBuffer::addDelta() became a lot more expensive, even though that code
wasn't changed at all. The reason is likely again cache effects: to be able to
reuse the zeros I had to allocate the buffer on the heap instead of on the
stack. Seemingly this bigger memory footprint has a big impact.