Bit 6 of internal register is "don't care" on MSX1 VDP
Tested on real machines by hap and myself: SNOW26 should show grey
flashes. But these do not appear on V99x8.

I guess that the original 'fix' (commit 4ec80c17) was done because the
grey flashing was not seen on a real MSX2, and that was considered a
bug. But apparently the flashing is intentional and only seen on MSX1
VDP's.