Reduce size of VDP-access-slot lookup tables
The VDPAccessSlot code offers a function that can answer the following:
  When is the next access slot that is at least 'c' cycles past time 't'?
This function is needed each time the CPU or the command engine reads or writes
from/to VRAM. So potentially it's called thousands of times per second. These
access slots are unevenly spread over one display line (see doc/internal/
vdp-vram-timing for a lot more details). To speedup the calculation we use
various lookup tables (LUTs).

These LUTs have one entry per cycle in a display line (1368 cycles per line).
And they refer to another cycle in the line. So, before this patch, each LUT
had 1368 entries with each entry holding a 16-bit value (values in range
0-1368). Though if we store relative instead of absolute values in the LUT, so
the delta to the next slot instead of the absolute position of that slot, we
can fit each entry in only 8 bits.

So this halves the size of the LUTs. Reducing the total memory size from ~150kB
to ~75kB. In the future we will add a few more LUTs (for text/character modes),
so this becomes even more important. Smaller memory footprint should result in
more efficient cache usage, though I couldn't measure this effect (I saw a 0.2%
speedup, but that's comparable to the measurement noise). I should also add
that this new storage format does not require more complex calculations (in one
routine this generates one x86 instruction less, in another one instruction
more).