Remove 'userData' from Schedulable
Made Schedulable objects more light-weight by removing the 'userData' member.
And thus also remove the corresponding parameter from the executeUntil() and
related methods. Schedulable objects do get copied around quite a bit, so
reducing their size from 24 to 16 bytes (when taking alignment into account)
does make a (small) measurable difference.

This is a follow-up on the Scheduler refactorings I did half a year ago. I only
now found the time to finally finish it (the backwards-compatible savestate
part kept on pending). For reference see:
  cccafc7826: [1/3] Scheduler optimizations, pendingSyncpoint()

The 'userData' stuff was used when the same Schedulable class could get
scheduled for 2 or more different reasons. The Scheduler code itself doesn't
care about the userData stuff, it only passes it to the executeUntil() method.
The implementation of that method could then behave different for the various
sync-point types. Now that parameter is gone, so instead we need multiple
Schedulable objects. In c++ it's not possible to implement the same interface
multiple times, so instead of inheriting once from Schedulable, we now have
multiple sub-objects that each inherit from Schedulable. And these sub-objects
typically have a back-pointer to the original object.

When comparing both alternatives, for the case that actually used 'userData',
we replaced the userData with a back-pointer variable. That's not directly a
win (though the pointer is located in a part that doesn't get copied). But the
majority of Schedulables don't actually need 'userData' and there we gain more.