Optimize CPUClock::setLimit()
This method is called each time Scheduler::setSyncPoint() is caller. Thus each
time the CPU must stop potentially sooner than said before. As explained in
earlier commits, the Scheduler is now more intensely used than in the past, so
the setLimit() function became relatively more expensive.

The CPU emulation main-loop must be interrupted for 3 reasons:
 a) a synchronization point is reached
 b) a 'slow' CPU instruction must be executed
 c) at 'regular' time intervals
See comments in the code (e.g. CPUClock.hh) for lots more details. Before this
patch CPUClock::setLimit() was responsible for a) and c). This patch removes
responsibility c) because it was redundant.

In the current code, the actual emulation code (which is driven by CPU
emulation) runs on the main thread. There are various helper threads that
sometimes 'post' work that must also happen on the main thread. So at regular
time intervals the main thread must stop executing CPU emulation and perform
those other tasks. It's not important that those tasks happen immediately,
'soon' is good enough.

In the past processing of host input events was done (only) at the above
mentioned exit points. So for interactivity it was indeed important these
points occur very regularly. Now processing of input events is (also) tied to
the frame rendering, and that already guarantees smooth processing of the
input events.

To guarantee the CPU-loop is exited at regular intervals, before this patch,
the setLimit() function would set the CPU execution limit to maximally execute
for 15000 cycles (for Z80 that is ~4ms or ~240Hz). Though there are other
sources that already cause regular exits from the CPU loop:
 - The VDP sets 3 sync points per frame (not evenly spread, but max interval is
   approx 13ms)
 - MSXMixer causes exits at typically intervals of 44100Hz/512 (approx 12ms)
 - RealTime causes exits at intervals of 80ms.
 - Various other devices and events also cause exists, but those are less
   regular (and not always guaranteed).

So not enforcing these 15000 cycles only extends the max interval by a factor
3x (worst case). And as explained above the regularity of this interval is less
important than it used to be. So there's no longer a compelling reason to keep
it. Removing it results in a simpler setLimit() function that runs approx 20%
faster (0.3% total emulation time in some benchmarks).