Fixed regression in 5th-sprite status register
Revision 57fa3a9b (22 Nov 2007) did some preparations that allowed to greatly
speedup the checkSprites{1,2} routines. Unfortunately it also introduced a
regression: the 5th (or 9th) sprite-number on a line was no longer guaranteed
to be detected at the earliest display line.

The following BASIC program demonstrates the problem. It should print '201'
(verified on a real machine), but before this patch it wrongly printed '200'.

  100 COLOR15,1,1:SCREEN2,2
  110 FORI=0TO31:A$=A$+CHR$(255):NEXT
  120 SPRITE$(0)=A$
  130 PUTSPRITE0,( 0,50),2,0
  140 PUTSPRITE1,(20,51),3,0
  150 PUTSPRITE2,(40,52),4,0
  160 PUTSPRITE3,(60,53),5,0
  170 PUTSPRITE4,( 0,80),2,0
  180 PUTSPRITE5,(20,81),3,0
  190 PUTSPRITE6,(40,82),4,0
  200 PUTSPRITE7,(60,83),5,0
  210 PUTSPRITE9,(80,54),6,0    <-- this is the earliest 5th sprite
  220 PUTSPRITE8,(80,84),6,0    <-- but openMSX reported this one
  230 FORI=0TO100:NEXT
  240 A=VDP(8)
  250 SCREEN0
  260 PRINTA

Luckily it's easy to fix while at the same time keeping the improved
performance. Check the comments at the top of checkSprites1() for more details.

This patch also makes a few smaller changes (mostly cosmetic) to make the
different checkSprites{1,2} routines more similar.