Store spriteCount[] in 8 bit (instead of 32)
There can maximally be 32 sprites on a line (normally only 4 or 8, but with
limitsprites=off openMSX can show all 32 sprites on one line). So we can store
the elements of the spriteCount[] array in 8 bits instead of 32.

(Un)fortunately the array was not that big, only 313 entries, so this change
doesn't result in a measurable performance difference.