Optimize OSDWidget::paint{GL,SDL}Recursive()
I noticed that in some benchmarks the OSDWidget::paintSDLRecursive() method was
taking about 3% of total time, even though only one OSD element (the mute icon)
was visible. The reason why this was so slow turned out to be copying of
shared_ptr objects. Copying a shared_ptr requires atomic operations (for the
reference count, even though in this case the shared_ptr's weren't passed
between different threads). (On x86) an atomic increment/decrement is only one
instruction, but it executes very slowly (10-100 cycles)!

The shared_ptr's were copied for the following reason:

   While drawing OSD elements there can occur errors. In such case a CliComm
   warning is sent. We have Tcl scripts that listen for such warnings and show
   them using the OSD stuff. So that means that during drawing there can be new
   OSD elements created or existing ones destroyed (or modified).

   In C++ it's not allowed to iterate over a changing vector (in this case that
   means new widgets are created or existing ones are destroyed, or existing
   elements are reshuffled (because they are sorted on z-coordinate)). It's
   also not allowed to still use an already deleted object (in this case that
   means draw a just destroyed widget). To protect against both problems, the
   OSD rendering code would make a copy of the vector<shared_ptr<OSDWidget>>
   container and iterate over this copy. (Note we're only copying pointers not
   the whole widgets). The copy protects against the changing vector, and
   holding the widgets by shared_ptr extends their lifetime till after the draw
   loop has finished (till after the copy is discarded).

This patch solves the above problem in a different way: when an error is
detected during drawing, we queue that error, and only at the end we send
all warnings to CliComm. I measured that this makes the paintSDLRecursive()
function between 2x and 3x faster.