Final(?) tweaks to vdp-timing-v2 article
Let it rest for a week, re-read it, and made some minor additions/corrections.

The timing diagram also contained (minor) mistakes (spotted by grauw). I've
fixed them and re-generated the image. To reduce the filesize I've also ran
'pngquant' on it, this reduces the image to 8-bit paletted colors. (Next I ran
'optipng' on it, but that's something we already do for most (all?) of our
PNGs.)