[2/8] MSXMixer::generate(), stereo with balance=center
If a stereo SoundDevice (e.g. both moonsound parts) has balance=center, we can
perform the 'mul' or 'mulAcc' operation with only half the amount of
multiplications as the general case. And the simplified 'mul/mulAcc' operations
turn out to be exactly the same as we already have for the 'mono with
balance=center'-case, so we can reuse those.

It's almost always the case that moonsound has balance=center (this is not the
panning of the OPL4 chip itself, but the openMSX user overriding the balance
setting). So IMHO this extra special case is worth it.