[3/8] MSXMixer::generate(), generate/mul in-place
This patch moves the call to updateBuffer() from one common location to each of
the 8 specialized locations (these cases are mono/stereo, balance=center/
not-center and first/not-first). This allows to also specialize the parameters
passed to this call for all these 8 cases.

The latter case (first or non-first) interests us in this patch. For the first
non-silent sound device we only need to multiply the generated samples, for all
later devices we also need to accumulate the samples with those of the earlier
devices. This patch goes one step further: for the first device we can generate
the samples directly into the destination buffer (the still empty accumulation
buffer). So we pass a different pointer to updateBuffer().

It's very often the case that MSX software only uses 1 sound device, so this
special case triggers very often. And also when the software uses 2 devices
(e.g. MSX-MUSIC for background music, PSG for sound effects) this case still
triggers half of the time. And when it triggers it halves the amount of touched
memory which is good for cache-efficiency (and this is becoming more and more
important).