[4/8] MSXMixer::generate(), only check samples==0 once
The previous patch also duplicated the 'samples==0'-test 8 times. And the same
condition was already present in some other locations. The code becomes simpler
if we handle this case once at the beginning of the function.

Background info:
'samples==0' means we're asked to generate zero host samples (typically host
sample rate is 44100Hz). So at first sight samples==0 means we shouldn't do
anything (and that's what we did some years ago). Though number of host samples
is rounded to an integer amount. It's possible that before rounding e.g. 0.5
host samples have passed since the last call to MSXMixer::generate() (or better
said: a time that corresponds to 0.5 host samples). The sample rate of the MSX
sound devices is not necessarily the same as the host sample rate. So it's
possible that although zero host samples must be generated, there _has_ passed
enough time to have a non-zero amount of msx-device-samples. For sound quality,
especially when playing Z80-driven samples (e.g. on PSG or key-click), it's
very important to also emulate the emulated sound devices at a per-sample
granularity. So even when samples==0 we must still call the sound device
emulation code.