[6/8] MSXMixer::generate(), overlap monoBuf and output
Before this patch MSXMixer::generate() used the following buffers:
 monoBuf   : sum of all mono   msx device samples
 stereoBuf : sum of all stereo msx device samples
 tmpBuf    : temp output of SoundDevice::updateBuffer() is written here
             before it's multiply-accumulated in monoBuf/stereoBuf (but see
             an earlier patch for when we can skip this buffer)
 output    : final result

The first 3 buffers hold 32-bit values, 'output' is 16-bit. 'output' is also
'stereo'. So 'monoBuf' and 'output' have the same size (1 channel 32-bit vs 2
channels 16-bit). It also turns out that the lifetime of both buffers does not
overlap (the last read from 'monoBuf' comes before the first write to 'output',
at least when you look at a per-sample granularity). So we can reuse the same
memory for both buffers. And this has a positive effect on cache-efficiency
(and as said before, cache-efficiency is becoming more and more important).

Combined with the earlier patches in this series, this divides the memory usage
by 3 for the common case (only 1 mono sound device active). And this is indeed
a very common scenario. (But also the general case has become faster).