[5/8] MSXMixer::generate(), use int16_t/int32_t
The MSXMixer::generate() deals with 16-bit and 32-bit samples. Internally all
processing is done on 32-bits. This allows to add multiple channels together
and multiply them by an 8-bit (later 9-bit) scaling factor without risk of
overflow. Only at the end we convert the result to 16-bit (and possibly we
clip at [-32768,32767]).

This patch uses the types 'int16_t' and 'int32_t' (instead of 'short' and
'int') to make it clearer where which type of sample(buffer) is used. Apart
from these renames there should be no changes.