[7/8] MSXMixer::generate(), tweaks for auto-vectorization
This patch tweaks some code so that both gcc and clang can do a better job at
auto-vectorization (using SSE2 instructions).

Two very often used sub routines are mul() and mulAcc(). We already have ARM
asm versions of these routines. One property that's used in those asm routines
is to process the samples in groups of 4, and if the amount of to-be-processed
samples is not an integer multiple of 4, we round it up. So we potentially
process up to 3 samples too many. But that's OK, the code can handle this.

This is also a useful property when working with SSE2 instructions (one SSE
register is 128-bit or 4x32-bit integers). It turns out that both gcc and clang
can generate much cleaner vector code when they only need to deal with groups
of 4 samples.

SSE instructions are faster on 16-bit aligned memory (but this becomes less
important on newer x86 CPUs). In an earlier patch I already made sure our
buffers are 16-byte aligned. This patch adds a new utility 'assume_aligned'
that allows the compiler to assume the buffers are indeed properly aligned (ATM
only gcc understands this hint).