Fix CPU type autodetection on little endian MIPS (mipsel)
Linux Python will return "mips" instead of "mipsel"; "arch" returns
"mips" as well. So if "mips" is reported, we'll ask the native compiler
for its machine identifier.

While we have compiled for mipsel a lot, most of that was cross
compilation, in which case CPU autodetection is not used. The Debian
builds however are native, which means that they have been broken for
the last few years. The most obvious symtom is that all SHA1 sums
mismatch with those found in hardware configs and software DB.