Don't use Scheduler to control INTRQ in WD2793
'INTRQ' is a status bit of the WD2793. It's set/reset depending on various
stuff. One of the possibilities is that it's set after an index pulse is
received (when the floppy has reached a certain rotation angle). We can
calculate the moment in (emu)time this condition will occur. Before this patch
we would set a synchronization point at that moment in time, and when it
triggers we change the INTRQ boolean variable.

Though if all we do in the synchronization point routine is setting a boolean
we can also use the following approach:
 - replace 'bool INTRQ' with 'EmuTime irqTime'
 - (current-time >= irqTime) <=> (INTRQ == true)
So if in the old code we want to make INTRQ=false, we now set irqTime=infinity.
If we want INTRQ=true, we set irqTime=zero. And if we want to make INTRQ true
at some point 't' in the future, we set irqTime=t.

This patch by itself is a small optimization. Though the main reason for this
patch is as a preparation step towards a more light-weight Scheduler class. I'm
investigating whether it's possible to get rid of the 'userData' parameter (see
Scheduler code). This patch already removes it from the WD2793 code.