Implement VRAM access slots for character and text modes
Details are in doc/internal/vdp-vram-timing/vdp-timing-2.html.
This only adds the V99x8 access slots. TMS9918 is still TODO.

Before this patch bitmap, character and text modes behaved the same (we only
made a distinction for screen on/off, sprites on/off). If you compare bitmap
and character mode, the access slot tables are _very_ close (but not
identical). So I doubt you'll see any difference in character mode after this
patch. However the table for the text modes is quite different.