[8/8] MSXMixer::generate(), remove >>8 from DC filter
After all the previous tweaks, the most expensive remaining part of
MSXMixer::generate() is the DC-removal filter at the end. We want such a filter
because:
 - Some devices only output positive values, and that halves the dynamic range
   of the signal.
 - Some sound devices sometimes keep their output at a constant non-zero
   value (not good for some audio equipment, also reduces dynamic range)
Anyway, this patch speeds up that filter in a small way.

Our current DC filter uses a 1st order IIR filter:
  y(n) = x(n) - x(n-1) + R * y(n-1)
with
  x(n) the input stream, y(n) the output stream, R a constant.
So the new output value depends on the new input value, the previous input
value and the previously calculated output value, more on this below. For the
constant R we took the value 511/512 (see comments in source for more details).

Right before the DC filter the audio signals are multiplied (and accumulated)
to apply the volume and balance settings. Because the signal is a stream of
integers this multiplication is done in 24.8 fixed-point. E.g. to multiply by
1.5 we actually multiply by 384 followed by a division by 256 (actually shift
right by 8 instead of a real division, also more on this below).

So both the volume/balance- and the DC-filter-step have a division operation.
If we use 23.9 bit fixed point then both are even divisions by 512 and the above
formula becomes:
  y(n) = (x(n) - x(n-1) + 511 * y(n-1)) / 512
So both divisions can be folded into one.

As already mentioned, instead of '/ 256' we did a '>> 8' operation, but that is
_not_ the same: division rounds towards zero, shift rounds down, so different
behavior for negative values. If we want the DC-filter to eventually fall to
zero (on constant input) it's required to round towards zero. For the volume/
balance calculation rounding isn't critical, but it's nicer to also round towards
zero. Another small bonus is that we have one bit more for the multiplication
factors and that the intermediate result isn't rounded before filtering.

--

This already is a long commit message for a relatively small patch. But I also
did some research that in the end didn't end up in this patch (it didn't give
better results). I'd still like to very briefly mention it.

In a IIR filter the new output depends on the previously calculated output.
This makes it hard to calculate different output samples in parallel, and I
guess this is (part of) why this routine runs relatively slow on modern x86
CPUs. Normally a x86 CPU works on many instructions in parallel, even
out-of-order and even across multiple loop iterations! (The latest CPUs have
scheduling windows of over 100 instructions!) But when most instructions depend
on the result of an earlier instruction this doesn't bring much benefit. That's
why I looked at a few alternative DC-filter implementations.

An obvious choice is a FIR filter (output depends only on the N-last input
samples). Though to get the same quality (low ripple and narrow transition
band) as the IIR filter, you'd have to use a too high order filter (N very
high). This might be feasible if we could integrate the FIR filter in another
FIR filter in the pipeline. Our HQ resampler uses a FIR filter. Though that
resampler is optional. (I mean both because it's conditionally enabled via
the resampler setting and because it's skipped if input and output sample
rate are already the same (e.g. OPL4 wave part)).

Another alternative is to subtract a moving average from the signal. This
requires only a few additions/subtractions per sample, but to efficiently
calculate the average it also needs a buffer of the last so many samples
(approx 2000 samples if we want to keep frequencies above 20Hz intact). This
large buffer might be slow because of cache effects. Though more important: I
looked at the spectra of an impulse response of this filter and audio-quality-
wise it's much worse than the IIR filter. I found a paper that drastically
improves the quality, but at the cost of 4 moving averages (and thus 4x the
buffer space).

While looking at the spectra of the impulse responses, I found that they indeed
match the theory (e.g. with R=511/512 @44100HZ our IIR filter has indeed a -3dB
attenuation at 14Hz). But only when using floating point calculations! When
using integer operations I saw much worse results. Though I'm not sure that
looking at the impulse response is the best analysis tool in this case. If
openMSX only had to run on PC, we should consider using floating point for the
full-audio-pipeline (on modern x86 CPUs, floating point operations are almost
as fast as integer operations).