Fix(?) bug#282 Y8950 is incorrectly handling key-off for percussive tones
This makes the rhythm key-on/off handling the same as the other OPLx variants.
I also verified it on a real Y8950 using the following BASIC program. This
program sets up the registers to allow to play various rhythm sounds (lines
500-570). Lines 140-160 intentionally wrongly set the keyon bit in channels
7-9. Before this patch the rhythm sounds were audible. After this patch, just
like on a real Y8950, they are not.

	10 READR,D
	20 IFR=0THENEND
	25 IFR=255THENGOSUB91
	30 OUT&HC0,R:OUT&HC1,D
	40 GOTO10
	91 FORI=0TO100:NEXT:RETURN
	100 DATA &hbd,&h00
	110 DATA &ha6,&h80
	120 DATA &ha7,&h60
	130 DATA &ha8,&h13
	140 DATA &hb6,&h20 :'
	150 DATA &hb7,&h20 :'
	160 DATA &hb8,&h30 :'
	170 DATA &h30,&h02
	180 DATA &h50,&h00
	190 DATA &h70,&hf8
	200 DATA &h90,&h46
	210 DATA &h31,&h0e
	220 DATA &h51,&h00
	230 DATA &h71,&hf8
	240 DATA &h91,&h46
	250 DATA &h32,&h0f
	260 DATA &h52,&h00
	270 DATA &h72,&hf8
	280 DATA &h92,&h47
	290 DATA &h33,&h08
	300 DATA &h53,&h00
	310 DATA &h73,&hf8
	320 DATA &h93,&h46
	330 DATA &h34,&h00
	340 DATA &h54,&h00
	350 DATA &h74,&hf8
	360 DATA &h94,&h46
	370 DATA &h35,&h00
	380 DATA &h55,&h00
	390 DATA &h75,&hf8
	400 DATA &h95,&h45
	500 DATA 255,0
	510 DATA&hbd,&he8
	520 DATA 255,0
	530 DATA&hbd,&he0
	540 DATA 255,0
	550 DATA&hbd,&he8
	560 DATA 255,0
	570 DATA&hbd,&he0
	999 DATA 0,0

This should fix bug #282, though when testing the scenario described in that
bug report it behaves the same before and after this patch. But it's possible
I'm just doing something wrong (when replaying the scenario). And in any case,
this patch does fix a real bug.