Fix bug#282 Y8950 is incorrectly handling key-off for percussive tones (2nd attempt)
This should fix bug#282 for real. Now I do see a difference in behavior before
and after this patch in the scenario described in the bug report.

This makes the code that handles the percussive stuff the same in Y8950 (OPL1)
and YMF262 (OPL3). It's different for the YM2413 (OPLL) cores, but according to
FRS OPLL indeed handles this in a different way.