Replace Alarm with RTScheduler
The Alarm class was built around the SDL_Timer functionality. This would
execute a callback after a specific amount of (host) time has passed. A
disadvantage of the SDL_Timer is that the callback runs in a different than the
main thread. Often we actually want to do some action on the main thread.
Before this patch the solution was to only post an event from the callback
(posting events is thread-safe) and then in the main thread handle that event.
(This wasn't the case initially, but) after some years all of our
timer-callbacks followed this pattern of posting/handling events.

Though when all timer callbacks simply post an event, we can as well in the
main thread code that delivers the events check for expired timers. This has
a few advantages:
- Adding/removing/check-for-pending/check-for-expired timers all happens within
  one thread, there's no locking needed anymore (and the Alarm code did contain
  some tricky code to work around an SDL threading problem).
- At first sight it may seem that a callback is more responsive than events
  that are only delivered from time-to-time. Though when the callback posts an
  event you still have to wait for that event to be delivered, and you actually
  get the same (or even a slightly longer) total response time.
- In the last few weeks the (EmuTime) Scheduler code was optimized quite a bit.
  The same optimization can be applied to scheduling real-time events. And
  because those optimization are located in a separate SchedulerQueue class, we
  could almost direct reuse those optimizations.
- This new approach is closer to a typical event-loop you find in GUI
  frameworks like Qt.

When benchmarking this code I got mixed results:
- When using valgrind (callgrind) I got about the same result before and after
  this patch. This is what I expected.
- But when measuring via 'perf stat' I measured an 8%(!) speedup.

Usually 'perf' gives more accurate results, but I'm not sure I should fully
trust the result now (it looks too good). A partial explanation might be that
in valgrind the test runs close to 100% speed (maybe slightly less) while with
'perf stat' the test can run at 10x-15x real-time speed. When you run slower
real-time timers expire relatively more often compared to when you run faster.
So both test-methods didn't run exactly the same code.

Another part of the explanation could be that 'valgrind' cannot measure
inter-thread communication overhead (e.g. cache-bouncing), 'perf' uses the CPUs
performance counters, so it takes everything into account.

Anyway, I don't believe that in realistic scenarios (running at 100% speed
instead of as fast as possible) you'll see 8% less CPU usage (maybe 1% or
less). But even if there's no performance difference at all, this patch is
still nice because it simplifies the code quite a bit.