Inline trivial getter/setters
I noticed we had quite a lot of trivial single line methods that were
implemented in the .cc file. Often the overhead of calling such a method is
more than simply inlining that body (not just the call/return itself, but more
important the optimizations that are prevented because of calling an arbitrary
function).

To create this patch I grepped in all .cc files for one-line methods that were
a single return statement or a single assignment. When it was really a trivial
non-virtual method (and when it didn't require moving #includes from .cc to
.hh) I moved the implementation of that method to the header file.

None of these methods were used in performance critical code. But I did measure
the executable size got 8kB-12kB smaller after this patch (depending on the
build flavour). Also the code got about 700 lines shorter.