Restore console reaction time in pause mode
When openMSX is paused (via 'set pause on' or via 'debug break'), we would only
process the event loop 10x per second. In the past that was fine, because as
soon as an input event was received, we broke out of the
EventDistributor::sleep() call and processed the event right away. And polling
for events was done asynchronously at least 40x per seconds (more if emulation
was running).

After revision 7bd9111ba polling for events is done synchronously (simplifies
stuff a lot), so the break-out-of-sleep stuff doesn't work anymore. Fixed by
polling more often.