Don't update reversebar when it's invisible
When running benchmarks I saw the StrimgMapImpl::findKey() function relatively
high in the profile results (taking 2% of total CPU time). This function is
used when looking up a string in a (hash)map. I didn't expect string operations
to take a noticeable amount of time during normal openMSX operation.

It turned out this function was (indirectly) called from the Tcl script that
draws the reversebar. (Mostly I run benchmarks with reverse disabled, because
that's still the typical situation on hand-held systems). The script is
constantly reconfiguring the OSD elements that make up the reversebar, each
element is looked up by name.

I didn't see an obvious way to speedup the lookup. But very often the
reversebar is invisible, so most of those update aren't really needed. This
patch extends the script to detect when the reversebar is completely faded away
and in that case skip the updates.