Replace CliComm libxml input parser with ad-hoc parser
The goal of this patch is to get rid of the dependency on libxml. We only
used this library (anymore) for parsing the stream of incoming commands
send over a CliComm connection (e.g. via stdin or via a socket). Catapult
and the openMSX-debugger both use this.

The input side of this connection is very simple:
  <openmsx-control>
    <command>my command</command>
    <command>another command</command>
    ...
So one opening 'openmsx-control' tag (typically there's no corresponding
closing tag because we're sending a whole stream). And each command is enclosed
in a <command> tag. There are no attributes.

For such a simple format it's relatively easy to write a custom parser. The
parser in this patch does not accept full-xml anymore. E.g. it cannot handle
attributes, it cannot handle nested tags, it ignores the opening tag, but it
can handle xml-entities like '&lt;'. See the (commented out) unittest in the
code for examples of what it can and cannot handle. I've tested that both
catapult and openMSX-debugger still work fine.

This patch changes the parser, but it does not yet update the build system
to actually remove libxml. That's for a later patch.