Cache sha1 in File class
While running 'perf top' we noticed that FilePool::findInDatabase() was taking
relatively much time (almost 1%). The expectation was that during normal
emulation the FilePool code isn't used at all. Manuel figured out that this
only occurs when there's a disk image inserted and reverse is enabled.

Reverse takes regular snapshots of the whole machine state. We include the sha1
of the currently inserted disk images in the snapshots. This allows to give a
warning when we later reverse to a point in time where the disk image content
was different. To calculate the sha1 of a .dsk image we use the
File::getSha1Sum() method. This in turn uses the FilePool::getSha1Sum() method.

FilePool::getSha1Sum() uses an (in-memory) database of files with known sha1
(the .filecache stuff). If the file is present in the database (and the file-
modification timestamp check passes) we can directly return the sha1 without
having to (re)calculate it. For large files like disk images this can be a big
speedup. Though looking up sha1 by filename is not the primary use case of
FilePool, so it was implemented as a O(N) operation (with N the number of files
in the pool). And that's why it's relatively high in the profile.

One possible solution is to also add a reverse-map in the FilePool (to allow
O(log(N)) lookups in both directions). Though that would increase the cost of
all FilePool operations and almost double the memory consumption. The
alternative solution we take in this patch is to also cache the sha1 in the
File class. So the first getSha1Sum() is still O(N), but all later calls (for
as long as the file doesn't change) are O(1).